Created a Custom Hook with all the logic that was previously done in App.jsx and Header.jsx}
