+++ src/App.jsx
@@ -1,7 +1,6 @@
-import { useState, useEffect } from "react"
 import Header from "./components/Header"
 import Guitar from "./components/Guitar"
-import { db } from "./data/db"
+import { useCart } from "./hooks/useCart"
 
 function App() {
 
@@ -21,87 +20,21 @@ function App() {
     }, 3000);
     */
 
-    const initialCart = () => {
-        const localStorageCart = localStorage.getItem("cart");
-        return localStorageCart ? JSON.parse(localStorageCart) : [];
-    }
-
-    const [data, setData] = useState(db);
-    const [cart, setCart] = useState(initialCart);
-
-    const max_items = 5;
-    const min_items = 1;
-
-    useEffect(() => {
-        localStorage.setItem("cart", JSON.stringify(cart));
-    }, [cart])
-
-    function addToCart(item) {
-
-        const itemExists = cart.findIndex((guitar) => guitar.id === item.id);
-        //console.log(itemExists)
-
-        if(itemExists >= 0) {
-            if(cart[itemExists].quantity >= max_items) {
-                return;
-            }
-            //console.log("El Producto ya existe")
-            //cart[itemExists].quantity++; // This mutates the original array 
-
-            const updatedCart = [...cart];
-            updatedCart[itemExists].quantity++;
-            setCart(updatedCart);
-
-        } else {
-            //console.log("no existe, agregando...")
-            item.quantity = 1; // Add quantity
-            setCart([...cart, item]);
-        }
-
-        
-    }
-
-    function removeFromCart(id) {
-        setCart(prevCart => prevCart.filter(guitar => guitar.id !== id))
-    }
-
-    function increaseQuantity(id) {
-        const updatedCart = cart.map( item => {
-
-            if(item.id === id && item.quantity < max_items) {
-                return {
-                    ...item,
-                    quantity: item.quantity + 1
-                }
-            }
-
-            return item;
-        })
-        setCart(updatedCart);
-    }
-
-    function decreaseQuantity(id) {
-        const updatedCart = cart.map( item => {
-
-            if(item.id === id && item.quantity > min_items) {
-                return {
-                    ...item, 
-                    quantity: item.quantity - 1
-                }
-            }
-
-            return item;
-        })
-        setCart(updatedCart);
-    }
-
-    function clearCart() {
-        setCart([]);
-    }
-
     //function saveLocalStorage() {
     //    localStorage.setItem("cart", JSON.stringify(cart));
     //}
+
+    const { 
+        data, 
+        cart, 
+        addToCart, 
+        removeFromCart, 
+        increaseQuantity, 
+        decreaseQuantity, 
+        clearCart, 
+        isEmpty, 
+        cartTotal 
+    } = useCart();
 
     return (
             <>
@@ -111,6 +44,8 @@ function App() {
                 increaseQuantity={increaseQuantity}
                 decreaseQuantity={decreaseQuantity}
                 clearCart={clearCart}
+                isEmpty={isEmpty}
+                cartTotal={cartTotal}
             />
 
             <main className="container-xl mt-5">
@@ -121,7 +56,6 @@ function App() {
                             <Guitar 
                                 key = {guitar.id}
                                 guitar = {guitar}
-                                setCart={setCart}
                                 addToCart={addToCart}
                             />
                         )
